raspi/requirements.txt
eventlet
python-socketio